refactor: move subtractExcessPerPattern on utils & delete useless function
